+++ lib/flow.dart
@@ -2,3 +2,4 @@ library flow;
 export 'src/flowimpl.dart';
 export 'src/extensions.dart';
 export 'src/builders.dart';
+export 'src/cache.dart';