+++ lib/src/exceptions/cancellation_exception.dart
@@ -1,4 +1,0 @@
-class CancellationException implements Exception {
-  final Exception cause;
-  CancellationException(this.cause);
-}